Move dialog_manager request models to /requests.
