+++ lib/modules/dialog_manager/dialog_manager.dart
@@ -1,9 +1,9 @@
 library dialog_manager;
 
 export 'src/enums/confirm_dialog_response_type.dart';
-export 'src/models/confirm_dialog_request.dart';
-export 'src/models/custom_dialog_request.dart';
-export 'src/models/informative_dialog_request.dart';
+export 'src/models/requests/confirm_dialog_request.dart';
+export 'src/models/requests/custom_dialog_request.dart';
+export 'src/models/requests/informative_dialog_request.dart';
 export 'src/services/dialog_service.dart';
 export 'src/services/i_dialog_service.dart';
 export 'src/widgets/dialog_manager.dart';
