+++ lib/modules/dialog_manager/src/models/requests/base_dialog_request.dart
