+++ lib/modules/dialog_manager/src/models/requests/custom_dialog_request.dart
