+++ lib/modules/dialog_manager/src/models/requests/description_dialog_request.dart
