+++ lib/modules/dialog_manager/src/models/requests/informative_dialog_request.dart
