+++ lib/modules/dialog_manager/src/services/dialog_service.dart
@@ -3,9 +3,10 @@ import 'dart:async';
 import 'package:meta/meta.dart';
 
 import '../enums/confirm_dialog_response_type.dart';
-import '../models/base_dialog_request.dart';
-import '../models/confirm_dialog_request.dart';
-import '../models/informative_dialog_request.dart';
+import '../models/requests/base_dialog_request.dart';
+import '../models/requests/confirm_dialog_request.dart';
+import '../models/requests/custom_dialog_request.dart';
+import '../models/requests/informative_dialog_request.dart';
 import 'i_dialog_service.dart';
 
 class DialogService implements IDialogService {
