+++ lib/modules/dialog_manager/src/services/i_dialog_service.dart
@@ -1,9 +1,10 @@
 import 'package:meta/meta.dart';
 
 import '../enums/confirm_dialog_response_type.dart';
-import '../models/base_dialog_request.dart';
-import '../models/confirm_dialog_request.dart';
-import '../models/informative_dialog_request.dart';
+import '../models/requests/base_dialog_request.dart';
+import '../models/requests/confirm_dialog_request.dart';
+import '../models/requests/custom_dialog_request.dart';
+import '../models/requests/informative_dialog_request.dart';
 
 abstract class IDialogService {
   Stream<BaseDialogRequest> get requestStream;
