+++ test/modules/dialog_manager/services/dialog_service_test.dart
@@ -1,5 +1,5 @@
 import 'package:bratacha/modules/dialog_manager/dialog_manager.dart';
-import 'package:bratacha/modules/dialog_manager/src/models/base_dialog_request.dart';
+import 'package:bratacha/modules/dialog_manager/src/models/requests/base_dialog_request.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
